build: divide melhor as funcionalidades do display em .c e .h
